@@ -5,10 +5,8 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.app.entities.Language;
+import com.app.entities.City;
 @Repository 
-public interface LanguageRepository extends JpaRepository<Language , Long> {
-
-	List<Language> findByNameIn(List<String> languageNames);
-
+public interface CityRepository extends JpaRepository<City, Long> {
+    List<City> findByCountry_CountryId(Long countryId);
 }
