@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import com.app.entities.User;
-
+@Repository 
 public interface UserRepository extends JpaRepository<User, Long> {
 	
 	/**
